refactor: moved user route to separate file
