+++ backend/app/api.py
@@ -7,7 +7,8 @@ import os
 import openai
 from . import crud, models, schemas
 from .database import SessionLocal, engine
-from .routers import chatgpt, auth, file
+from .routers import chatgpt, auth, file, users
+from .middleware.authHandler import JWTBearer
 
 
 # Load environment variables
@@ -25,7 +26,8 @@ app = FastAPI()
 # Include routers
 app.include_router(chatgpt.router)
 app.include_router(auth.router)
-app.include_router(file.router)
+app.include_router(users.router)
+app.include_router(file.router, dependencies=[Depends(JWTBearer())])
 
 # Configure CORS middleware
 origins = ["http://localhost:3000", "localhost:3000"]
@@ -78,34 +80,3 @@ async def get_questions(question: Question) -> Response:
     )
     return {"response": response}
 
-# User routes
-@app.post("/users/", response_model=schemas.User)
-def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
-    db_user = crud.get_user_by_email(db, email=user.email)
-    if db_user:
-        raise HTTPException(status_code=400, detail="Email already registered")
-    return crud.create_user(db=db, user=user)
-
-@app.get("/users/", response_model=list[schemas.User])
-def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    users = crud.get_users(db, skip=skip, limit=limit)
-    return users
-
-@app.get("/users/{user_id}", response_model=schemas.User)
-def read_user(user_id: int, db: Session = Depends(get_db)):
-    db_user = crud.get_user(db, user_id=user_id)
-    if db_user is None:
-        raise HTTPException(status_code=404, detail="User not found")
-    return db_user
-
-@app.post("/users/{user_id}/items/", response_model=schemas.Item)
-def create_item_for_user(
-    user_id: int, item: schemas.ItemCreate, db: Session = Depends(get_db)
-):
-    return crud.create_user_item(db=db, item=item, user_id=user_id)
-
-# Read items route
-@app.get("/items/", response_model=list[schemas.Item])
-def read_items(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    items = crud.get_items(db, skip=skip, limit=limit)
-    return items
